Move all models to the model module
